Rename gedcomx files
